final_project/requirements.txt
boto3
modal
numpy
pandas
psycopg2
pytest
PyYAML
sentence_transformers
torch
tqdm
transformers
pytest-cov
